refactor(hosted): satisfy candidate quality gates
