+++ tests/private-hosted-cli/services-lifecycle.test.js
@@ -10,51 +10,6 @@ const { FakeWebSocket, settle } = require('../cluster/harness');
 const { makeAccess, waitForRequest, waitForSocket } = require('../hosted-session/harness');
 const { captureLogs, DESCRIPTOR } = require('./candidate-fixtures');
 const { remoteHarness } = require('./remote-service-harness');
-const { targetHarness } = require('./target-service-harness');
-
-const INTENT_ID = '019fd184-52c3-7e1f-a567-4ecb6fc6a0ec';
-const INTENT_CAPSULE_ID = '019fd184-58d2-7db4-a878-5bd495c986a4';
-const SUBMISSION_KEY = '019fd184-637d-4f26-af31-5ec3b3ef1dd6';
-const NOW = '2026-08-05T10:00:00.000Z';
-
-function intent(overrides = {}) {
-  return {
-    intent_id: INTENT_ID,
-    state: 'queued',
-    waiting_reason: null,
-    capsule_id: null,
-    result: null,
-    error_code: null,
-    submitted_at: NOW,
-    updated_at: NOW,
-    terminal_at: null,
-    ...overrides,
-  };
-}
-
-describe('private target services', () => {
-  it('runs add, login, list, setup, and remove through production service wiring', async () => {
-    const h = targetHarness();
-    await captureLogs(() => h.services.targetAdd('next', { url: 'https://target.example' }));
-    await captureLogs(() => h.services.targetLogin('prod'));
-    await captureLogs(() =>
-      h.services.targetSetup('prod', {
-        repository: 'owner/repository',
-        provider: 'codex',
-        modelLevel: 'level2',
-      })
-    );
-    const listed = await captureLogs(() => h.services.targetList({ json: true }));
-    assert.match(listed.lines[0], /"configured": true/);
-    await captureLogs(() => h.services.targetRemove('prod', { force: false }));
-
-    assert.deepEqual(h.state._targets.prod, undefined);
-    assert.equal(h.state._targets.next.id, 'target-next');
-    assert.equal(h.calls.filter(([name]) => name === 'login').length, 1);
-    assert.equal(h.calls.filter(([name]) => name === 'revoke').length, 1);
-    assert.equal(h.calls.filter(([name]) => name === 'delete').length, 1);
-  });
-});
 
 function registerCapsuleServiceTests() {
   it('creates, lists, and host-terminates capsules with distinct operations', async () => {
@@ -315,201 +270,10 @@ function registerRemoteOperationTests() {
   });
 }
 
-function registerQueueSubmissionTests() {
-  it('submits and follows a credential-free v2 RunIntent through one client', async () => {
-    const queueCalls = [];
-    const h = remoteHarness({
-      createRunIntentClient: () => ({
-        submit(request) {
-          queueCalls.push(['submit', request]);
-          return intent();
-        },
-        get(id) {
-          queueCalls.push(['get', id]);
-          return intent({
-            state: 'succeeded',
-            capsule_id: INTENT_CAPSULE_ID,
-            result: { summary: 'done' },
-            terminal_at: NOW,
-          });
-        },
-      }),
-      runIntentSleep: () => Promise.resolve(),
-    });
-    const result = await captureLogs(() =>
-      h.services.remoteQueueRun({
-        target: 'prod',
-        graph: 'graph.json',
-        input: 'input.json',
-        queue: true,
-        submissionKey: SUBMISSION_KEY,
-        detach: false,
-      })
-    );
-    assert.equal(result.value.state, 'succeeded');
-    assert.deepEqual(
-      queueCalls.map(([name]) => name),
-      ['submit', 'get']
-    );
-    const submitted = queueCalls[0][1];
-    assert.equal(submitted.submissionKey, SUBMISSION_KEY);
-    assert.deepEqual(Object.keys(submitted.envelope), ['version', 'graph', 'input']);
-    assert.equal(submitted.envelope.version, 'zeroshot.run-intent/v2');
-    assert.deepEqual(submitted.envelope.input, {
-      source: 'prompt',
-      prompt: 'Ship the change.',
-      artifacts: [],
-    });
-  });
-
-  it('fails closed when the target does not advertise RunIntent v2', async () => {
-    const h = remoteHarness({
-      descriptor: { ...DESCRIPTOR, runIntent: null },
-    });
-    await assert.rejects(
-      h.services.remoteQueueRun({
-        target: 'prod',
-        graph: 'graph.json',
-        input: 'input.json',
-        queue: true,
-        submissionKey: SUBMISSION_KEY,
-        detach: true,
-      }),
-      /does not advertise RunIntent v2/
-    );
-  });
-}
-
-function registerQueueObservationTests() {
-  it('keeps Ctrl+C as queue observation disconnect without cancellation', async () => {
-    const queueCalls = [];
-    const h = remoteHarness({
-      createRunIntentClient: () => ({
-        submit() {
-          queueCalls.push(['submit']);
-          return intent();
-        },
-        get(_id, options) {
-          queueCalls.push(['get']);
-          process.emit('SIGINT');
-          return Promise.reject(options.signal.reason);
-        },
-        cancel() {
-          queueCalls.push(['cancel']);
-        },
-      }),
-      runIntentSleep: () => Promise.resolve(),
-    });
-    const listenersBefore = process.listenerCount('SIGINT');
-    const result = await captureLogs(() =>
-      h.services.remoteQueueRun({
-        target: 'prod',
-        graph: 'graph.json',
-        input: 'input.json',
-        queue: true,
-        submissionKey: SUBMISSION_KEY,
-        detach: false,
-      })
-    );
-    assert.equal(result.value.state, 'queued');
-    assert.equal(process.listenerCount('SIGINT'), listenersBefore);
-    assert.equal(
-      queueCalls.some(([name]) => name === 'cancel'),
-      false
-    );
-    assert.match(result.lines.join('\n'), /was not cancelled/);
-  });
-
-  it('preserves exact-key recovery guidance after an ambiguous submission', async () => {
-    let submissions = 0;
-    const h = remoteHarness({
-      createRunIntentClient: () => ({
-        submit() {
-          submissions += 1;
-          throw new Error('peer-secret-detail');
-        },
-      }),
-    });
-    await assert.rejects(
-      captureLogs(() =>
-        h.services.remoteQueueRun({
-          target: 'prod',
-          graph: 'graph.json',
-          input: 'input.json',
-          queue: true,
-          submissionKey: SUBMISSION_KEY,
-          detach: true,
-        })
-      ),
-      (error) => {
-        assert.match(error.message, new RegExp(`--submission-key ${SUBMISSION_KEY}`));
-        assert.doesNotMatch(error.message, /peer-secret-detail/);
-        return true;
-      }
-    );
-    assert.equal(submissions, 1);
-  });
-}
-
-function registerRunIntentManagementTests() {
-  it('reads status, follows, and cancels only through explicit RunIntent operations', async () => {
-    const queueCalls = [];
-    const h = remoteHarness({
-      createRunIntentClient: () => ({
-        get(id) {
-          queueCalls.push(['get', id]);
-          return intent({
-            state: 'succeeded',
-            capsule_id: INTENT_CAPSULE_ID,
-            result: { summary: 'done' },
-            terminal_at: NOW,
-          });
-        },
-        cancel(id) {
-          queueCalls.push(['cancel', id]);
-          return intent({ state: 'cancelling', capsule_id: INTENT_CAPSULE_ID });
-        },
-      }),
-    });
-    await captureLogs(() =>
-      h.services.runIntentStatus('prod', INTENT_ID, { follow: true, json: false })
-    );
-    await captureLogs(() => h.services.runIntentCancel('prod', INTENT_ID));
-    assert.deepEqual(queueCalls, [
-      ['get', INTENT_ID],
-      ['cancel', INTENT_ID],
-    ]);
-  });
-
-  it('reports a purged successful result without inventing an empty object', async () => {
-    const h = remoteHarness({
-      createRunIntentClient: () => ({
-        get() {
-          return intent({
-            state: 'succeeded',
-            capsule_id: INTENT_CAPSULE_ID,
-            result: null,
-            terminal_at: NOW,
-          });
-        },
-      }),
-    });
-    const result = await captureLogs(() =>
-      h.services.runIntentStatus('prod', INTENT_ID, { follow: true, json: false })
-    );
-    assert.equal(result.value.result, null);
-    assert.match(result.lines.join('\n'), /result is no longer retained/);
-    assert.doesNotMatch(result.lines.join('\n'), /^\{\}$/m);
-  });
-}
-
 function registerCapsuleAndObservationServiceTests() {
   registerCapsuleServiceTests();
   registerReconnectTest();
   registerRemoteOperationTests();
-  registerQueueSubmissionTests();
-  registerQueueObservationTests();
-  registerRunIntentManagementTests();
 }
 
 describe('private capsule and observation services', registerCapsuleAndObservationServiceTests);
